Move all commands in separate files
