+++ src/main/kotlin/ru/hse/sd/cli/lang/impl/CommandParser.kt
@@ -3,7 +3,7 @@ package ru.hse.sd.cli.lang.impl
 import com.github.h0tk3y.betterParse.lexer.TokenMatchesSequence
 import com.github.h0tk3y.betterParse.parser.parseToEnd
 import ru.hse.sd.cli.command.Command
-import ru.hse.sd.cli.command.EmptyCommand
+import ru.hse.sd.cli.command.impl.EmptyCommand
 import ru.hse.sd.cli.lang.Parser
 
 /**
