+++ src/test/kotlin/ru/hse/sd/cli/lang/impl/CommandParserTest.kt
@@ -4,6 +4,7 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
 import ru.hse.sd.cli.command.*
+import ru.hse.sd.cli.command.impl.*
 import kotlin.test.assertEquals
 
 internal class CommandParserTest {
